Cambio de layout de home, quitar logo de react y agregar logo de elecciones.
